+++ src/zodbbrowser/__init__.py
@@ -18,6 +18,5 @@ ZODB Browser has the following submodules:
 
 """
 
-
 __version__ = '0.12.1.dev0'
 __homepage__ = 'https://github.com/mgedmin/zodbbrowser'